app/models/Api.java

+++ app/models/Api.java
@@ -2,6 +2,7 @@ package models;
 
 import helpers.ControllerHelper;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
@@ -42,7 +43,7 @@ public class Api extends ApiModeloBase {
 	@Min(value = 0,message="error.precioMes.min")
 	private Integer precioMes;
 	
-	private Double puntuacionMedia;
+	private Float puntuacionMedia;
 	
 	@Required(message="error.obligatorio.xml")
 	private boolean xml;
@@ -161,11 +162,11 @@ public class Api extends ApiModeloBase {
 		this.precioMes = precioMes;
 	}
 
-	public Double getPuntuacionMedia() {
+	public Float getPuntuacionMedia() {
 		return puntuacionMedia;
 	}
 
-	public void setPuntuacionMedia(Double puntuacionMedia) {
+	public void setPuntuacionMedia(Float puntuacionMedia) {
 		this.puntuacionMedia = puntuacionMedia;
 	}
